+++ CHANGELOG.md
@@ -1,2 +1,2 @@
 ### 1.0.0
-- Initial package using the same as Custom-Code-Editor in ApostropheCMS version 2.x.x
+- Initial package using the same as Custom-Code-Editor in ApostropheCMS version 2.x.x. First Release for ApostropheCMS version 3.x.x!